Change to 1-based positions
While I personally like 0-based for positions, switching our schema to use them results in a ton of confusion when integrating with other code bases. I experienced this when integrating with dbsnp/snpedia. It's too easy to think a read means something different because of that small off by one error.