Update and rename referenceHierarchy.avdl to referenceAlignments.avdl
Sorry to hijack this file, I'm proposing to generalise it to represent alignments, with two possible constructs - the alignment block and a cigar alignment. The alignment block is a pairwise structure that matches intervals between two blocks, and can be placed in an array to describe a pairwise alignment. The second alignment structure is a CIGAR based alignment that apes the basic CIGAR format, with the difference that the alignment is between to paths in two sequence graphs.