Added "complete" field to allow incomplete records
For a SAM/BAM/CRAM without the SA tag, it is very difficult to know whether a
read has multiple linear alignments or to collect all the linear alignments of
the read. With a "complete" field, we allow to specify a GARead object not to
include all the linear alignments.

This commit also changes GASearchReadsResponse to return fragments. In addition,
it adds GASearchReadsRequest::separateFragment. When it is flagged, one
fragment may be returned in multiple GAFragment objects, mimicing typical BAM
queries.