make GAReferenceSequence::sequence required
We want the API to get the sequence, not the user.
And this simplifies a lot of the logic in the API.
The one thing I am not sure about is whether this means that in implementation when you get the ReferenceSequenceSet then you get all the literal reference sequences at the same time.  This would be ~3Gb for a human genome, so heavy.  More sophisticated implementations would get sequences only on demand, and cache.  I don't understand enough about the methods to know whether that is possible
with this schema.

Also rename uri to sourceURI.