Compose ReferenceSet.md5 from all Reference.md5
This was originally proposed on ga4gh#66 by @vadimzalunin. The advantage of this approach is that computation of ReferenceSet.md5checksum is strictly easier than the original definition. Furthermore, it should have indistinguishable uniqueness guarantees.

Examples of implementation difficulties with the original approach:
- To create a new ReferenceSet, I need to fetch all sequence data contained within all references. The md5s are more likely to be readily available e.g. in the SAM header.
- To create a new ReferenceSet and corresponding References, I would need to compute md5 sums over all sequence data twice.
- To create a new ReferenceSet which uses existing References, but differs only in order from an existing ReferenceSet, I would need to compute and md5 over all Reference.sequence again.

Also makes Reference.name non-nullable, as the semantics of performing an md5 over a null reference name are/were not defined.