Update of metadata.avdl
This is a re-edition of the metadata.avdl change regarding the GASex definition, competing with @cassiedoll's changes in https://github.com/ga4gh/schemas/pull/138 
It implements some of the ideas/wishes from the discussion there and adds addtl. fixes:

* type "OTHER"
* explanation of definition/use
* deletes the "NOT_AVAILABLE" type since base is NA => 'null'
* removes the "GA" prefix

However, the move to common is not implemented here & will be done later if enough agreement.