Variation Reference Methods Adding a draft of some methods for the variation reference methods. There's a lot of work to do to sync up the representations of samples with the current variants API.
Substantially revising variationReference.avdl to mirror some of the newer stuff from the Variants and References APIs. Not yet updating the methods.

First pass at updating the methods

Adding dataset enumeration and improving some comments.

Adding Back Alleles and Segments

After talking with Benedict, it makes sense to me what these are for and why we need them. We need to be able to make calls on portions of variants, so we can call the bit of a variant between two other variants on it, for example.
And if we can do that, we really ought to be able to call several of these Segments at once as one Allele (or so Benedict says). And then we ought to be able to call both ends of, say, an inversion together as one Allele, without reading through the whole thing.

And then we need the Scaffolds back to read through individual Alleles and not multi-Allele-at-once Calls. Abd they need to be associated with call sets and variant sets.

Updating functions to support Alleles and AlleleScaffolds

Rearranging some fields on variants and tweaking search.

Using arrays and and-of-ors semantics for all search options in the Variation Reference API

Fixing incorrect method name

Better name for name substring search field

Pluralizing some fields that were singular

Calls aren't actually on AlleleScaffolds anymore

Making sure we can search calls by allele

Adding a way to walk down the variant graph