Changing available search options.
Now it is possible to look up `Variant`s by `Alele`s they use.

I've also defined "simple" and "composite" `Allele`s for those that are defined as `Path`s versus those that are defined as sets of other "child" `Allele`s, and specified when composite `Allele`s ought to match a search.