Refactoring the way sequence graphs are represented.
We have a set of sequences, which are linear, possibly empty DNA strings with
IDs.

These sequences are joined at their ends, and we can use a set of `Segment`s
each covering a whole sequence to describe the resulting sequence graph.

We can also describe a path through the graph as an ordered list of
`Segment`s.

`VariantSet`s can report `Segment`s describing the new pieces of sequence
graph with which to augment the `ReferenceSet`'s sequence graph before
interpreting the `Allele`s.