Bringing graph support to Common, Reads, and References
There is now a `Path` and `Segment` in a graph namespace in common.avdl, and a
`GraphAlignment` in the graph namespace in reads.avdl. We also have a
`/graph/sequence/{id}` endpoint to get sequence bases in a generic way, and a
couple more options to search by.

Adding support for monoallelic calls.

This adds an `Allele` type, and allows for `Call`s to be associated with
`Allele`s, or with `Allele`s and `Variant`s.

`Allele`s can be used to represent the ref and alt alleles of `Varaint`s.

We also provide for associating new sequences with `VariantSet`s.

Removing graph namespace and URL prefix

Revising comments and field names

Unifying checksums

Revising variants comments

Code reviewed by Prof. Haussler

Changing comments to be more descriptive and mathematically tight.

Also catching a few typos.

Fixing oversights in new API calls

Adding in primer on sequence graphs

Revising David's sequence graph intro

Explaining sequence search better

Sequences are searched over, and results are returned as segments describing
sequences.

Also removing references to an out-of-date object name.

Removing update timestamp from Allele

Clarifying sequences, segments, and references

Fixing up whitespace

Introducing "graph" and "classic" modes.

A server can support "graph" mode, "classic" mode, or both.

If it supports "classic" mode, it will make sure to fill in all of
the fields that non-graph clients require, and not use incompatible
graph extensions like `GraphAlignment`.

If it supports "graph" mode, it will make sure to fill in all of the new
graph fields, like `Reference.segment` and `Variant.alleles`. Additionally,
if it also does not support "classic" mode, it will not use older, less
general things like `LinearAlignment` and single string phase sets.

Making David's documentation changes

Also adding more pictures and hammering on the parent/child distinction.

Making startJoin and endJoin default to null

Clarifying the order of segments in a path

Fixing example typos

Splitting out `AlleleCall`.

There is now an `AlleleCall` to take care of calling copy numbers of `Alleles`,
while the normal `Call` goes back to just handling genotypes. I also stripped
some of the bachwards-compatibility warts off of `AlleleCall`, since it has no
legacy fields to keep.

Both kinds of calls are constrained to be consistent with each other; you
can't call a genotype of [1, 1], and have a copy number of 0 for that allele
in that variant.

If you support "graph" mode, you have to serve `AlleleCall`s for every
`Variant` you have `Call`s on, and visa versa if you support "classic" mode.

More descriptive name for new method

Fixing case where it would be impossible to get allele sequence.

Since a `Position` can be specified by `Reference` name, there is no
efficient way to search for a `Reference` by name, and the only way to get
sequence bases (in the old and proposed APIs) is by ID, we need to be able
to make sure that all `Position`s include IDs in the graph mode, so we can
actually get the sequence.

Fixing doc comment typos

Clarifying computation of `Reference.md5checksum`.

I now give an explicit procedure for how you are supposed to compute the
`md5checksum` of a `Reference`. It no longer makes reference to the idea of a
checksum for a `Position`, since that doesn't actually exist.

It should be much clearer now what is to be hashed, which is important because
everyone needs to do the same thing to get the same hashes.