Updating methods to work with side graphs.
Only allowing one range to be passed at a time in `SearchJoinsRequest`.

Also, adding in `extractSubgraphSegments()` and `extractSubgraphJoins()`
to get the subgraph you touch if you start at a position and go out a
certain number of bases.

Fixing an incorrect default value

Forcing server deduplication.

Fixing a spelling error