Making `Alleles` first-class objects.
`Allele`s now exist outside of, and can be shared between,
`VariantSet`s. The parent object of an `Allele` is a `Dataset`.

Adding a `searchAllelesPath()` method that can use any of a number of
"algorithms" to get `Allele`s that are "relevant" to a `Path`, for some
algorithm-defined notion of relevance.

Also specifying that the `Allele`s used by a `VariantSet` may only use a
"simple" extension of the `Sequence`s and `Join`s in the relevant
`ReferenceSet`. That is, no novel `Join` may have both ends in novel
`Sequence`s.