metadata.avdl time formats
This PR addresses the previously discussed use of ISO8601 for date/time attributes. One problem here is the lack of a specific dateTime format in the current AVRO implementation (this will be addressed in v1.8(?)). For now, the PR clarifies the ISO use instead of the vague "timestamp" description, and recommends the proper formatting with the attribute type set to string.

Other changes :

.created => . recordCreated
.updated => recordUpdated
Experiment.runDate => Experiment.runTime
Sample.age => Sample.ageAtSampling