Rationalises the SearchObjectRequest semantics.
Change the semantics of all SearchObjectRequests so they all require
the ID of a single container object, if applicable. Currently, a
number of different conventions exist, making implementation and
authorisation difficult.

The sole exception to this rule is the SearchReadsRequest, which
allows us to search over a list of readGroups. This allows us to
merge the reads over multiple readGroups efficiently.