Update metadata.avdl
This PR introduces  ```Dataset``` record type (at the expense of the previous limited ```IndividualGroup```). Please see https://github.com/ga4gh/schemas/issues/248 for some background on this.
```Dataset``` is seen with inherent *evolvability*. There was consensus among the metadata task team that the best way would be to put this forward in a skeleton format with future feature refinement.