replace CallSet id and name with label
This replaces the id (server-owned) and name (user-owned) for a CallSet by a label (user-defined, subject to uniqueness and syntax constraints).  This allows user-meaningful labels for the equivalent of VCF columns, without semantic problems of clashes.

Also it moves the info map for a call set to a new CallMethod object, with a methodId in the call set, and a default one in the variant set.
This allows sharing of meta-information about the calling process and other aspects of call sets.

Other changes are that I made call-sets only belong to one variant set (why could they belong to many?) and removed the updated timestamp because currently we have no way to update.

The new concept is Label, which is a string of printable characters, unique within the scope of the variant set, possibly with restricted syntax e.g. [a-zA-Z][a-zA-Z0-9_]* like programming language identifiers.  The aim is that this should be a reasonable thing for users to look at and use, e.g. to display as column headings, or use in queries.
This replaces both columnId, which is server-unique and not under user control, and name, which is user-defined and can’t be relied on to be unique.
We are OK to require the labels to be unique, because the scope is limited to the VariantSet, so long range clashes are not a problem.  The semantics of merging VariantSets can be specified as part of the merge operation - both options of maintaining uniqueness by adding suffixes if necessary or forcing merge when the labels coincide have merit in different circumstances.
Because the labels must be unique, they can be used as a key for retrieval of CallSet columns in the VariantSet.
Having them separate from the sampleId means that you can have two CallSet columns for the same sample in one VariantSet, which is nice to be able to compare call sets made in different ways from the same sample (e.g. Illumina versus CG, or different callers).
Having a methodId that references an external Method object, which contains an info map for metadata, allows columns to share the same information about how they were called.  We also add an optional methodId to the VariantSet itself, which would then act as a default for all its callsets, covering the standard case where all callsets are made the same way.

It is good for both sampleId and methodId to be optional, so that lightweight VariantSets can be made by importing VCF files or the equivalent, without having to create lots of other empty objects first.  We should not underestimate the importance of lightweight use of the object representation and API - a large amount of sequence data handling is done in small labs with LIMS that will autopopulate Sample objects, or in exploratory analysis, and we would like users not to have to manage unnecessary appendages.

I need to add changes to variantMethods to this pull request, but don't know how to just now in this web interface.