Update instructions for documentation generation
We change the documentation readme and add messages to a few places to
point would-be documentation generators towards running `mvn package` in
the repository root. We also let them know where the generated
documentation will land. Finally, we add links in the READMEs to
`INSTALL.md` and to the ReadTheDocs documentation, so they don't need to
build it themselves at all.