Revise Avro and JSON documentation
In GA4GH world, we use only the AVDL pnd JSON parts of Avro, and we
don't use (for example) Avro RPC. This previously was not documented.
Now it is.

To make the Avro JSON libraries work, there ae some restrictions on
legal unions. I describe those, and how, because of this restriction,
Avro JSON just looks like normal JSON.

I've also put some information (to be expanded) on how we define an API
by defining a bunch of objects.