Add scripts for releasing to Maven Central (resolves #278)
Adds a directory `scripts/release` with a script `scripts/release/release.sh`
that contains scripts for pushing a release of the GA4GH schemas to Maven
Central. These scripts will build the GA4GH schema artifacts, sign them, and
push them to the Sonatype Nexus OSS staging repository. From there, the
artifacts are uploaded to Maven Central. Additionally, a tag will be created
and pushed to the Github schemas repository.