Change ontology term identifier (#694)
* Change ontology term identifier
Ontology terms had an ambigious 'id' field and description. This PR renames the field to term_id to remove the confusion over the role of the field, as well as improves the comment for its proper use. A field 'term_uri' has been added that allows one to specify the full URI for a term if it is available.
Change ontologies appendix documentation to reflect change
Slight formatting and wording adjustments of appendix doc
Change CURIE to URI
Fix casing

* post Vancouver changes

modified after Vancouver meeting to use CURIE throughout, remove
termURI and remove sourceName and source version.

* fixed termId

* doc

* Fix casing

* Update process schemas